requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
openpyxl>=3.1.0  # For Excel file support

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality (optional)
black>=23.0.0
flake8>=6.0.0

# Jupyter support
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Visualization
matplotlib>=3.7.0

# Optional LLM integration (Gemini) and env support
google-generativeai>=0.8.0
python-dotenv>=1.0.0
